+++ output/models/nist_data/list_pkg/double/schema_instance/nistschema_sv_iv_list_double_max_length_1_xsd/nistschema_sv_iv_list_double_max_length_1.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from decimal import Decimal
 from typing import List
 
 __NAMESPACE__ = "NISTSchema-SV-IV-list-double-maxLength-1-NS"
@@ -11,7 +10,7 @@ class NistschemaSvIvListDoubleMaxLength1:
         name = "NISTSchema-SV-IV-list-double-maxLength-1"
         namespace = "NISTSchema-SV-IV-list-double-maxLength-1-NS"
 
-    value: List[Decimal] = field(
+    value: List[float] = field(
         default_factory=list,
         metadata={
             "required": True,
